IsingGenerator {bootnet}	R Documentation

Generates a function that simulates data from the Ising model

Description

Uses IsingSampler to generate the data.

Usage

IsingGenerator(...)

Arguments

...
Arguments passed to IsingSampler

Value

A function with as first argument the sample size and as second argument a named list, with element graph encoding a weights matrix and element intercepts encoding thresholds.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package bootnet version 1.1.0 Index]