genGGM {bootnet}	R Documentation

Generates a GGM small-world network.

Description

Simulates a GGM as described by Yin and Li (2011), using the Watts and Strogatz (1998) algorithm for generating the graph structure (see watts.strogatz.game).

Usage

genGGM(Nvar, p = 0, nei = 1, parRange = c(0.5, 1), constant = 1.5, propPositive = 0.5)

Arguments

Nvar
Number of nodes

p
Rewiring probability (see watts.strogatz.game).

nei
Neighborhood (see watts.strogatz.game).

parRange
Range of partial correlation coefficients to be originally sampled.

constant
A constant as described by Yin and Li (2011).

propPositive
Proportion of edges to be set positive.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

References

Yin, J., and Li, H. (2011). A sparse conditional gaussian graphical model for analysis of genetical genomics data. The annals of applied statistics, 5(4), 2630.

Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of ‘small-world’ networks. nature, 393(6684), 440-442.

[Package bootnet version 1.1.0 Index]