ggmGenerator {bootnet}	R Documentation

Generates a function that simulates data from the Gaussian graphical model (GGM)

Description

Generates data given a partial correlation network. Data can be made ordinal by using a threshold model with equally spaced thresholds.

Usage

ggmGenerator(ordinal = FALSE, nLevels = 4)

Arguments

ordinal
Logical, should ordinal data be generated?

nLevels
Number of levels used in ordinal data.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package bootnet version 1.1.0 Index]