plot.netSimulator {bootnet}	R Documentation

netSimulator S3 methods

NA
plot.replicationSimulator
Description

Plot, print and summary methods for netSimulator output.

Usage

## S3 method for class 'netSimulator'
plot(x, xvar = "factor(nCases)",
                              yvar = c("sensitivity", "specificity", "correlation"),
                              xfacet = "measure", yfacet = ".", color = NULL,
                             ylim = c(0,1), print = TRUE,  xlab = "Number of cases", 
                              ylab, ...)
## S3 method for class 'netSimulator'
print(x, digits = 2, ...)
## S3 method for class 'netSimulator'
summary(object, digits = 2, ...)

## S3 method for class 'replicationSimulator'
plot(x, xvar = "factor(nCases)",
                              yvar = c("correlation","jaccard","replicatedEdges",
                              "replicatedZeroes"),
                              xfacet = "measure", yfacet = ".", color = NULL,
                             ylim = c(0,1), print = TRUE,  xlab = "Number of cases", 
                              ylab, ...)
## S3 method for class 'replicationSimulator'
print(x, digits = 2, ...)
## S3 method for class 'replicationSimulator'
summary(object, digits = 2, ...)

Arguments

x
Output of netSimulator.

object
Output of netSimulator.

xvar
String indicating the variable to be used on the x-axis.

yvar
String vector indicating the variable(s) to be used on the y-axis.

xfacet
String indicating the variable to be used on the horizontal facets (or "." to omit).

yfacet
String indicating the variable to be used on the vertical facets (or "." to omit).

color
String indicating the variable to be used in coloring boxplots.

ylim
Y-axis limits.

print
Logical, should the plot be printed? This helps printing the plots to PDF files.

digits
Number of digits to be used in print and summary method.

xlab
X-axis label

ylab
Y-axis label. Defaults to combining the values in yvar. Is hidden when xfacet = "measure", as then it is clear what the y-axis represent from the facet labels.

...
Not used.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package bootnet version 1.1.0 Index]