plot.bootnetResult {bootnet}	R Documentation

Plot method for bootnetResult objects

Description

Plots the graph using the qgraph package and the qgraph function. Defined as qgraph::qgraph(x[['graph']],labels=x[['labels']],...)

Usage

## S3 method for class 'bootnetResult'
plot(x, weighted, signed, directed, labels, 
          layout = "spring", parallelEdge = TRUE, cut = 0, 
          theme = "colorblind", ...)

Arguments

x
A bootnetResult object

weighted
Logical, should the analyzed network be weighted?

signed
Logical, should the analyzed network be signed?

directed
Logical, is the analyzed network directed? Usually does not have to be set and is detected automatically.

labels
Labels of the nodes. Defaults to the column names of the data if missing.

layout
Placement of the nodes. See qgraph. Always defaults to "spring".

parallelEdge
Should edges in directed networks be plotted parallel? See qgraph. Defaults to TRUE instead of FALSE (as in qgraph).

cut
Should scaling in width and saturation of edges be split? See qgraph. Defaults to 0 to disable cut (qgraph chooses a cutoff with at least 20 nodes).

theme
Theme of the edge and node colors. See qgraph. Defaults to "colorblind" rather than the default used in qgraph ("classic").

...
Arguments sent to qgraph.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package bootnet version 1.1.0 Index]