summary.bootnet {bootnet}	R Documentation

Summarize bootnet results

Description

Creates a data frame (wrapped as tbl_df) containing summarized results of the bootstraps.

Usage

## S3 method for class 'bootnet'
summary(object, statistics = c("edge", "intercept", "strength", 
                            "closeness", "betweenness", "distance"), perNode = FALSE,
                            rank = FALSE, ...)

Arguments

object
A bootnet object

statistics
The types of statistics to include in the summary table

perNode
Logical, should centrality estimates per node be plotted instead of correlation with original parameter. Only used in node and person drop bootstrap.

rank
Logical, should plots show rank of statistics instead of statistics?

...
Not used.

Value

A tbl_df (data frame) containing summarized statistics.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package bootnet version 1.1.0 Index]